Added files from my platformer game.
The game is created with HTML, JavaScript, and CSS. There are 3 viscous, bloodthirsty dragons that need to be defeated. The player plays as Greg the Hero in this action-packed platformer. The player will play through 3 unique levels, where they will have to dodge obstacles such as pits and spikes. The dragons will be shooting fireballs at the player while they are attempting to defeat them.